Extract methods from BlkDevicesTable to a reusable mixin
